user/drfx/requirements.txt
appdirs==1.4.3
asgiref==3.2.3
async-generator==1.10
attrs==19.3.0
automium==0.2.6
automium-web==0.1.1
bcrypt==3.1.7
bleach==3.1.4
cffi==1.14.0
Click==7.0
codegen==1.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
Django==2.2.3
django-crispy-forms==1.9.0
django-phone-field==1.8.0
djangorestframework==3.11.0
entrypoints==0.3
filelock==3.0.12
Flask==1.1.2
ipython-genutils==0.2.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.14.1
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
jupyterlab-pygments==0.1.1
legacy==0.1.6
login==0.0.6
MarkupSafe==1.1.1
mistune==0.8.4
nbclient==0.2.0
-e git+https://github.com/jupyter/nbconvert.git@3bd3c7f9c208b79f03a09f999828dfd139630038#egg=nbconvert
nbformat==5.0.6
nest-asyncio==1.3.2
networkx==2.4
numpy==1.18.3
pandas==1.0.3
pandocfilters==1.4.2
patsy==0.5.1
patterns==0.3
Pillow==7.0.0
pycparser==2.20
Pygments==2.6.1
pyparsing==2.4.6
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2019.1
pywin32==227
PyYAML==5.3.1
pyzmq==19.0.0
scipy==1.4.1
-e git://github.com/jonashagstedt/django-simple-signup.git@62ce3a4bf81d02ad122422c936901ed6275b658f#egg=simple_signup
six==1.14.0
sqlparse==0.3.0
statsmodels==0.11.1
svgwrite==1.3.1
testpath==0.4.4
tornado==6.0.4
tqdm==4.45.0
traitlets==4.3.3
Tree==0.2.4
virtualenv==20.0.18
webencodings==0.5.1
Werkzeug==1.0.1
